spring-recipie-app/src/main/java/guru/springframework/converters/UnitOfMeasureToUnitOfMeasureCommand.java
/**
 * 
 */
package guru.springframework.converters;

import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

import guru.springframework.commands.UnitOfMeasureCommand;
import guru.springframework.domain.UnitOfMeasure;
import lombok.Synchronized;

/**
 * @author Maor Zaken
 * Created on 26 Jan 2020
 */
@Component
public class UnitOfMeasureToUnitOfMeasureCommand implements Converter<UnitOfMeasure, UnitOfMeasureCommand> {

	@Synchronized
	@Nullable
	@Override
	public UnitOfMeasureCommand convert(UnitOfMeasure source) {
		if (source == null) {
			return null;
		}
		UnitOfMeasureCommand uomCommand = new UnitOfMeasureCommand();
		uomCommand.setDescription(source.getDescription());
		uomCommand.setId(source.getId());
		
		return uomCommand;
	}

}
